LiquidElectrolytes.jl

This package is in pre-release state. There are two ways using this package:

Via github URL:

pkg> add https://github.com/j-fu/LiquidElectrolytes.jl

Versions 0.0.x are registered in https://github.com/j-fu/PackageNursery. If you trust this registry, you can issue once

pkg> registry add https://github.com/j-fu/PackageNursery

and add the package via

pkg> add LiquidElectrolytes

Data of Electrolyte

LiquidElectrolytes.ElectrolyteData — Type

mutable struct ElectrolyteData <: AbstractElectrolyteData

Data for electrolyte.

nc::Int64

Number of ionic species. Default: 2

iϕ::Int64

Potential index in species list. Default: nc + 1

ip::Int64

Pressure index in species list Default: nc + 2

T::Float64

Temperature Default: (273.15 + 25) * K

D::Vector{Float64}

Mobility coefficient Default: fill((2.0e-9 * m ^ 2) / s, nc)

M0::Float64

Molar weight of solvent Default: (18.0153g) / mol

M::Vector{Float64}

Molar weight of ions Default: fill(M0, nc)

z::Vector{Int64}

Charge numbers of ions Default: [(-1) ^ (i - 1) for i = 1:nc]

Z::Vector{Float64}

Charge numbers scaled by F/RT Default: (z .* F) / (R * T)

c_bulk::Vector{Float64}

Bulk ion concentrations Default: fill((0.1mol) / dm ^ 3, nc)

ϕ_bulk::Float64

Bulk voltage Default: 0.0

p_bulk::Float64

Bulk pressure Default: 0.0

Γ_bulk::Int64

Bulk boundary number Default: 1

v0::Float64

Molar volume of solvent Default: 1 / ((55.4mol) / dm ^ 3)

κ::Vector{Float64}

Solvation numbers Default: fill(0, nc)

v::Vector{Float64}

Molar volumes of ions Default: fill(v0, nc)

ε::Float64

Dielectric permittivity of solvent Default: 78.49

pscale::Float64

Pressure scaling factor Default: 1.0e9

neutralflag::Bool

Electroneutrality switch Default: false

scheme::Symbol

Electroneutrality switch Default: :μex

logreg::Float64

Logarithm regularization Default: 1.0e-18

LiquidElectrolytes.AbstractElectrolyteData — Type

abstract type AbstractElectrolyteData

Abstract super type for electrolytes

LiquidElectrolytes.Cdl0 — Method

Cdl0(electrolyte)

Double layer capacitance at zero voltage for symmetric binary electrolyte.

Example

@siunits mol dm μF cm
ely=ElectrolyteData(c_bulk=fill(0.01*mol/dm^3,2))
round(Cdl0(ely),digits=4)/(μF/cm^2)
# output

22.85

LiquidElectrolytes.ldebye — Method

ldebye(electrolyte)

Debye length.

LiquidElectrolytes.chemical_potentials! — Function

chemical_potentials!(μ,u,electrolyte::AbstractElectrolyteData)

Calculate chemical potentials from concentrations.

Input:

μ: mutated, allocated memory for result

u: concentrations

Returns μ0, μ: chemical potential of solvent and chemical potentials of ions.

LiquidElectrolytes.rrate — Function

rrate(R0,β,A)

Thermodynamic reaction rate expression

rrate(R0,β,A)=R0*(exp(-β*A) - exp((1-β)*A))

Internal API

LiquidElectrolytes.charge — Function

charge(c,electrolyte)

Calculate charge from vector of concentrations

LiquidElectrolytes.vrel — Function

vrel(ic,electrolyte)

$v_{i,rel}=κ_i+\frac{v_i}{v_0}$

LiquidElectrolytes.c0_barc — Function

c0_barc(u,electrolyte)

Calculate solvent concentration $c_0$ and summary concentration $\bar c$ from vector of concentrations c using the incompressibility constraint (assuming $κ_0=0$):

\[ \sum_{i=0}^N c_i (v_i + κ_iv_0) =1\]

This gives

\[ c_0v_0=1-\sum_{i=1}^N c_i (v_i+ κ_iv_0)\]

\[c_0= 1/v_0 - \sum_{i=1}^N c_i(\frac{v_i}{v_0}+κ)\]

Then we can calculate

\[ \bar c= \sum_{i=0}^N c_i\]

LiquidElectrolytes.rlog — Function

rlog(u, electrolyte)

Regularized logarithm:

   rlog(u,electrolyte)= log(u+electrolyte.logreg)

LiquidElectrolytes.wnorm — Function

wnorm(u,w,p)

Weighted norm with respect to columns

Physical unit handling

Physical unit handling is based on Unitful.jl and PhysicalContstants.jl, prospectively this will be factorized out into another package. The rationale of these macros is the fact that Unitful.jl facilitates working with strongly typed values. Slightly counterintuitively to me, units with different prefixes are different types. Moreover parts of the Julia infrastructure is not (yet) written in Julia, prominently so sparse matrix solvers, and these are not able to handle units.

So we (mis)use the unit packages as databases and allow to retrive numerical values in SI base units

Missing docstring.

Missing docstring for @si_str. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @siunits. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @phconstants. Check Documenter's build log for details.