General remarks

All physical quantities are assumed to be consistently represented through their values expressed in basic SI units (m, kg, s, A, K, mol, cd), supported by the LessUnitful.jl package built on top of Unitful.jl.

Electrolyte data

LiquidElectrolytes.AbstractElectrolyteData — Type

abstract type AbstractElectrolyteData

Abstract super type for electrolytes.

LiquidElectrolytes.ElectrolyteData — Type

mutable struct ElectrolyteData{Tγ, Tcache, Texp, Tlog, Tflux} <: AbstractElectrolyteData

Data for electrolyte. It is defined using Base.@kwdef allowing for keyword constructors like

    ElectrolyteData(nc=3,z=[-1,2,1])

To see default values, just create an instance as ElectrolyteData(). Fields (reserved fields are modified by some algorithms):

nc::Int64: Number of ionic species.

na::Int64: Number of surface species

iϕ::Int64: Potential index in species list.

ip::Int64: Pressure index in species list

D::Vector{Float64}: Mobility coefficient

z::Vector{Int64}: Charge numbers of ions

M0::Float64: Molar weight of solvent

M::Vector{Float64}: Molar weights of ions

v0::Float64: Molar volume of solvent

v::Vector{Float64}: Molar volumes of ions

κ::Vector{Float64}: Solvation numbers of ions

c_bulk::Vector{Float64}: Bulk ion concentrations

actcoeff!::Any: actcoeff!(γ, c, p, ::ElectrolyteData) Activity coefficient function. Write activity coefficients into γ. Default: DGL_gamma!.

ϕ_bulk::Float64: Bulk voltage

p_bulk::Float64: Bulk pressure

Γ_bulk::Int64: Bulk boundary number

Γ_we::Int64: Working electrode boundary number

T::Float64: Temperature

ε::Float64: Dielectric permittivity of solvent

rexp::Any: Regularized exponential, default: exp (unregularized)

rlog::Any: Regularized logarithm, default: log (unregularized)

pscale::Float64: Pressure scaling factor. Default: 1.0e9

eneutral::Bool: Local electroneutrality switch. Default: false

upwindflux!::Any: Upwind flux caculation method for ionic species. This allows to choose between

μex_flux! (default, strongly preferrable): excess chemical potential (SEDAN) scheme

act_flux!: scheme based on reciprocal activity coefficients

cent_flux!: central scheme

weights::Vector{Float64}: Species weights for norms in solver control.

solvepressure::Bool: Solve for pressure.

This is true by default. Setting this to false can serve two purposes:

Use the pressure from the solution of a flow equation

Ignore the pressure contribution to the excess chemical potential

F::Float64: Faraday constant (fixed)

ε_0::Float64: Dielectric permittivity of vacuum (fixed)

RT::Float64: Molar gas constant scaled with temperature (derived)

Mrel::Vector{Float64}: Solvated molar mass ratio (derived)

barv::Vector{Float64}: Solvated molar volume (derived)

tildev::Vector{Float64}: Pressure relevant volume (derived)

γ_bulk::Vector{Float64}: Activity coefficient of at bulk interface (derived)

γk_cache::Any: Cache for activity coefficient calculation (reserved)

γl_cache::Any: Cache for activity coefficient calculation (reserved)

ϕ_we::Float64: Working electrode voltage (reserved) Used by sweep algorithms to pass boundary data.

edgevelocity::Union{Float64, Vector{Float64}}: Edge velocity projection (reserved).

scheme::Symbol: Scheme parameter. Deprecated and disabled. Use upwindflux! instead to change the discretization scheme.

LiquidElectrolytes.update_derived! — Function

update_derived!(electrolyte::ElectrolyteData)

Update derived electrolyte data.

The default values for electrolyte data are those of an symmetric 0.1M aqueous binary electrolyte at 298.5K with solvation number κ=10, ion molar volumes and masses similar to those of water molecules and diffusion coefficients 2.0e-9 $m^2/s$. All values given in SI base units:

using LiquidElectrolytes
ElectrolyteData()

ElectrolyteData{typeof(LiquidElectrolytes.DGL_gamma!), PreallocationTools.DiffCache{Vector{Float64}, Vector{Float64}}, typeof(exp), typeof(log), typeof(LiquidElectrolytes.μex_flux!)}(2, 0, 3, 4, [2.0e-9, 2.0e-9], [1, -1], 0.0180153, [0.0180153, 0.0180153], 1.805054151624549e-5, [1.805054151624549e-5, 1.805054151624549e-5], [10.0, 10.0], [100.0, 100.0], LiquidElectrolytes.DGL_gamma!, 0.0, 0.0, 2, 1, 298.15, 78.49, exp, log, 1.0e9, false, LiquidElectrolytes.μex_flux!, [1.805054151624549e-5, 1.805054151624549e-5, 1.0, 0.0], true, 96485.33212331001, 8.8541878128e-12, 2478.957029602388, [11.0, 11.0], [0.00019855595667870037, 0.00019855595667870037], [0.0, 0.0], [1.0, 1.0], PreallocationTools.DiffCache{Vector{Float64}, Vector{Float64}}([0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], Any[]), PreallocationTools.DiffCache{Vector{Float64}, Vector{Float64}}([0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], Any[]), 0.0, 0.0, :deprecated)

LiquidElectrolytes.dlcap0 — Method

dlcap0(electrolyte)

Double layer capacitance at zero voltage for symmetric binary electrolyte.

Example

using LessUnitful
ely = ElectrolyteData(c_bulk=fill(0.01ufac"mol/dm^3",2))
round(dlcap0(ely),sigdigits=5) |> u"μF/cm^2"
# output

22.847 μF cm^-2

LiquidElectrolytes.debyelength — Method

debyelength(electrolyte)

Debye length.

using LessUnitful
ely = ElectrolyteData(c_bulk=fill(0.01ufac"mol/dm^3",2))
round(debyelength(ely),sigdigits=5) |> u"nm"
# output

4.3018 nm

LiquidElectrolytes.chargedensity — Function

chargedensity(c,electrolyte)

Calculate charge density from vector of concentrations (in one grid point).

chargedensity(sol,electrolyte)

Calculate charge density vector from solution (on whole grid)

chargedensity(tsol,electrolyte)

Calculate charge densities from from time/voltage dependent solution solution (on whole grid)

LiquidElectrolytes.chemical_potential — Function

 chemical_potential(c, barc, p, barv, electrolyte)

Calculate chemical potential of species with concentration c

\[ μ = \bar v(p-p_{ref}) + RT\log \frac{c}{\bar c}\]

LiquidElectrolytes.chemical_potentials! — Function

chemical_potentials!(μ,u,electrolyte)

Calculate chemical potentials from concentrations.

Input:

μ: memory for result (will be filled)

u: local solution vector (concentrations, potential, pressure)

Returns μ0, μ: chemical potential of solvent and chemical potentials of ions.

using LessUnitful
ely = ElectrolyteData(c_bulk = fill(0.01ufac"mol/dm^3", 2))
μ0, μ = chemical_potentials!([0.0, 0.0], vcat(ely.c_bulk, [0, 0]), ely)
round(μ0, sigdigits = 5), round.(μ, sigdigits = 5)
# output

(-0.89834, [-21359.0, -21359.0])

Breaking change between v0.2 and 0.3

This function has been corrected in version 0.3. Before it used the molar volume $v$ instead of the effective molar volume $\bar v = v + κv_0$ to calculate the ion chemical potentials. Examples with κ=0 are not affected.

LiquidElectrolytes.rrate — Function

rrate(R0,β,A)
rrate(R0,β,A, electrolyte)

Reaction rate expression

rrate(R0,β,A)=R0*(exp(-β*A) - exp((1-β)*A))

LiquidElectrolytes.iselectroneutral — Function

iselectroneutral(cx::Vector,celldata)

Check for electroneutrality of concentration vector

iselectroneutral(tsol::TransientSolution,celldata)

Check for electroneutrality of transient solution

LiquidElectrolytes.isincompressible — Function

isincompressible(cx::Vector,celldata)

Check if the concentration vector fulfills incompressibility constraint, including the fact that the solvent concentration is nonnegative

isincompressible(tsol::TransientSolution,celldata)

Check for incompressibility of transient solution

LiquidElectrolytes.c0_barc — Function

c0_barc(u,electrolyte)

Calculate solvent concentration $c_0$ and summary concentration $\bar c$ from vector of concentrations c using the incompressibility constraint (assuming $κ_0=0$):

\[ \sum_{i=0}^N c_i (v_i + κ_iv_0) =1\]

This gives

\[ c_0v_0=1-\sum_{i=1}^N c_i (v_i+ κ_iv_0)\]

\[c_0= 1/v_0 - \sum_{i=1}^N c_i(\frac{v_i}{v_0}+κ_i)\]

Then we can calculate

\[ \bar c= \sum_{i=0}^N c_i\]

Poisson-Nernst-Planck system

LiquidElectrolytes.AbstractElectrochemicalSystem — Type

abstract type AbstractElectrochemicalSystem

Abstract super type for electrochemical systems

LiquidElectrolytes.PNPSystem — Type

PNPSystem(grid;
         celldata=ElectrolyteData(),
         bcondition=(f, u, n, e)-> nothing,
         reaction=(f, u, n, e)-> nothing,
         kwargs...)

Create VoronoiFVM system for generalized Poisson-Nernst-Planck. Input:

grid: discretization grid

celldata: composite struct containing electrolyte data

bcondition: boundary condition

reaction : reactions of the bulk species

kwargs: Keyword arguments of VoronoiFVM.System

LiquidElectrolytes.DGL_gamma! — Function

DGL_gamma!(γ, c, p, electrolyte)

Activity coefficients according to Dreyer, Guhlke, Landstorfer.

LiquidElectrolytes.electrolytedata — Function

electrolytedata(sys)

Extract electrolyte data from system.

LiquidElectrolytes.solventconcentration — Function

   solventconcentration(U::Array, electrolyte)

Calculate vector of solvent concentrations from solution array.

LiquidElectrolytes.chemical_potentials — Function

chemical_potentials(solution, electrolyte)

Calculate chemical potentials from solution. Returns μ0, μ, where μ0 is the vector of solvent chemical potentials, and μ is the nc×nnodes matrix of solute chemical potentials.

LiquidElectrolytes.electrochemical_potentials — Function

electrochemical_potentials(solution, electrolyte)

Calculate electrochemical potentials from solution. and return nc×nnodes matrix of solute electrochemical potentials measured in Volts.

Upwind fluxes

LiquidElectrolytes.μex_flux! — Function

μex_flux!(f,dϕ,ck,cl,γk,γl,electrolyte; evelo=0.0)

Excess chemical potential (Sedan) upweind flux, based on modified Scharfetter-Gummel scheme, see Gaudeul/Fuhrmann 2022.

Appearantly first described by Yu, Zhiping and Dutton, Robert, SEDAN III, www-tcad.stanford.edu/tcad/programs/sedan3.html

see also the 198? Fortran code available via https://web.archive.org/web/20210518233152/http://www-tcad.stanford.edu/tcad/programs/oldftpable.html

Verification calculation is in the paper.

LiquidElectrolytes.act_flux! — Function

act_flux!(ic,dϕ,ck,cl,γk,γl,electrolyte; evelo=0)

Scharfetter-Gummel upwind flux expression based on activities, see Fuhrmann, CPC 2015. As shown in Chainais-Hilliaret, Cances, Fuhrmann, Gaudeul, 2019, this is consistent to thermodynamic equilibrium but shows inferior convergence behavior.

LiquidElectrolytes.cent_flux! — Function

cent_flux!(ic,dϕ,ck,cl,γk,γl,electrolyte; evelo = 0)

Flux expression based on central differences, see Gaudeul/Fuhrmann 2022. As shown in the paper this is covergent, consistent to thermodynamic equilibrium but may show inferior convergence behavior.

Poisson-Boltzmann system

LiquidElectrolytes.PBSystem — Type

PBSystem(grid;
             celldata=ElectrolyteData(),
         bcondition=(f, u, n, e)-> nothing,
         kwargs...)

Create VoronoiFVM system generalized Poisson-Boltzmann. Input:

grid: discretization grid

celldata: composite struct containing electrolyte data

bcondition: boundary condition

kwargs: Keyword arguments of VoronoiFVM.System

Poisson-Nernst-Planck-Stokes system

LiquidElectrolytes.PNPStokesSolver — Type

struct PNPStokesSolver{Tflow, Tpnp}

Stokes solver struct.

flowsolver::Any

pnpsolver::Any

Utilities

LiquidElectrolytes.RLog — Type

RLog(eps)

Callable struct for regularized logarithm. Smooth linear continuation for x<eps. This means we can calculate a "logarithm" of a small negative number. Objects mylog::RLog of this type are meant to replace the logarithm function and allow to invoke mylog(x).

LiquidElectrolytes.RExp — Type

RExp(trunc)

Callable struct for regularized exponential. Linear continuation for x>trunc, returns 1/rexp(-x) for x<-trunc. Objects myexp::RExp of this type are meant to replace the exponential function and allow to invoke myexp(x).