Standard calculations

LiquidElectrolytes.bulkbcondition — Function

bulkbcondition(f,u,bnode,electrolyte)

Bulk boundary condition for electrolyte: set potential, pressure and concentrations to bulk values.

LiquidElectrolytes.AbstractSimulationResult — Type

abstract type AbstractSimulationResult

Abstract simulation result.

LiquidElectrolytes.voltages_solutions — Function

voltages_solutions(result)

Return a TransientSolution tsol containing voltages (in tsol.t) and the corresponding stationary solutions (in tsol.u).

Double layer capacitance

LiquidElectrolytes.dlcapsweep — Function

       dlcapsweep(sys;voltages=(-1:0.1:1)*ufac"V",
                              δ=1.0e-4,
                              molarity=0.1*ufac"mol/dm^3",
                              store_solutions=false,
                              solver_kwargs...)

Calculate double layer capacitances for voltages given in voltages. Returns a DLCapSweepResult

Assumptions:

Only one double layer in the system - close to working electrode

1D domain

LiquidElectrolytes.DLCapSweepResult — Type

struct DLCapSweepResult{Tv, Tc, Ts} <: AbstractSimulationResult

Result data type for dlcapsweep

voltages::Any: Vector of voltages

dlcaps::Any: Vector of double layer capacitances

solutions::Any: Vector of solutions

LiquidElectrolytes.voltages_dlcaps — Function

voltages_dlcaps(result)

Double layer capacitance curve as DiffEqArray

Voltage-current sweeps

LiquidElectrolytes.ivsweep — Function

 ivsweep(
      sys;
      voltages = (-0.5:0.1:0.5) * ufac"V",
      store_solutions = false,
      solver_kwargs...,
      )

Calculate molar reaction rates and bulk flux rates for each voltage in voltages.

LiquidElectrolytes.IVSweepResult — Type

struct IVSweepResult{Tv, Twe, Tbulk, Ts} <: AbstractSimulationResult

Result data type for ivsweep

voltages::Any: Vector of voltages

j_we::Any: Working electrode molar reaction rates

j_bulk::Any: Bulk molar fluxes

solutions::Any: Vector of solutions

LiquidElectrolytes.voltages_currents — Function

voltages_currents(result,ispec)

Voltage- working electrode current curve for species as DiffEqArray

LiquidElectrolytes.currents — Function

currents(result,ispec)

Electrode current for species ispec.