Data of Electrolyte

LiquidElectrolytes.ElectrolyteData — Type

mutable struct ElectrolyteData <: AbstractElectrolyteData

Data for electrolyte.

nc::Int64

Number of ionic species. Default: 2

iϕ::Int64

Potential index in species list. Default: nc + 1

ip::Int64

Pressure index in species list Default: nc + 2

T::Float64

Temperature Default: (273.15 + 25) * ufac"K"

D::Vector{Float64}

Mobility coefficient Default: fill(2.0e-9 * ufac"m^2/s", nc)

M0::Float64

Molar weight of solvent Default: 18.0153 * ufac"g/mol"

M::Vector{Float64}

Molar weight of ions Default: fill(M0, nc)

z::Vector{Int64}

Charge numbers of ions Default: [(-1) ^ (i - 1) for i = 1:nc]

RT::Float64

RT Default: ph"R" * T

F::Float64

Faraday constant Default: ph"N_A*e"

c_bulk::Vector{Float64}

Bulk ion concentrations Default: fill(0.1 * ufac"M", nc)

ϕ_bulk::Float64

Bulk voltage Default: 0.0 * ufac"V"

p_bulk::Float64

Bulk pressure Default: 0.0 * ufac"Pa"

Γ_bulk::Int64

Bulk boundary number Default: 1

v0::Float64

Molar volume of solvent Default: 1 / (55.4 * ufac"M")

κ::Vector{Float64}

Solvation numbers Default: fill(10.0, nc)

v::Vector{Float64}

Molar volumes of ions Default: fill(v0, nc)

ε::Float64

Dielectric permittivity of solvent Default: 78.49

ε_0::Float64

Dielectric permittivity of vacuum Default: ufac"ε0"

pscale::Float64

Pressure scaling factor Default: 1.0e9

neutralflag::Bool

Electroneutrality switch Default: false

scheme::Symbol

Electroneutrality switch Default: :μex

epsreg::Float64

Regularization parameter Default: 1.0e-20

LiquidElectrolytes.AbstractElectrolyteData — Type

abstract type AbstractElectrolyteData

Abstract super type for electrolytes

LiquidElectrolytes.Cdl0 — Method

Cdl0(electrolyte)

Double layer capacitance at zero voltage for symmetric binary electrolyte.

Example

using LessUnitful
ely=ElectrolyteData(c_bulk=fill(0.01ufac"mol/dm^3"),2))
round(Cdl0(ely)|>u"μF/cm^2",digits=2)
# output

22.85 μF cm^-2

LiquidElectrolytes.ldebye — Method

ldebye(electrolyte)

Debye length.

LiquidElectrolytes.chemical_potentials! — Function

chemical_potentials!(μ,u,electrolyte)

Calculate chemical potentials from concentrations.

Input:

μ: memory for result (will be filled)

u: concentrations

Returns μ0, μ: chemical potential of solvent and chemical potentials of ions.

LiquidElectrolytes.rrate — Function

rrate(R0,β,A)

Reaction rate expression

rrate(R0,β,A)=R0*(exp(-β*A) - exp((1-β)*A))

Discretization system

LiquidElectrolytes.PNPSystem — Function

PNPSystem(grid;
         celldata=ElectrolyteData(),
         bcondition=default_bcondition,
         kwargs...)

Create VoronoiFVM system. Input:

grid: discretization grid

celldata: composite struct containing electrolyte data

bcondition: boundary condition

kwargs: Keyword arguments of VoronoiFVM.System

LiquidElectrolytes.pnpunknowns — Function

pnpunknowns(sys)

Return vector of unknowns initialized with bulk data.

LiquidElectrolytes.electrolytedata — Function

electrolytedata(sys)

Extract electrolyte data from system.

LiquidElectrolytes.bulkbc — Function

default_bcondition(f,u,bnode,electrolyte)

Bulk boundary condition for electrolyte: set potential, pressure and concentrations to bulk values.

LiquidElectrolytes.c0 — Function

   c0(U::Array, electrolyte)

Calculate vector of solvent concentrations from solution array.

Standard calculations

LiquidElectrolytes.doublelayercap — Function

       doublelayercap(sys;voltages=(-1:0.1:1)*ufac"V",
                          δ=1.0e-4,
                          molarity=0.1*ufac"mol/dm^3",
                          solver_kwargs...)

Calculate double layer capacitances for voltages given in voltages. Returns vector of voltages and vector of double layer capacitances.

LiquidElectrolytes.voltagesweep — Function

voltagesweep(sys;
          voltages=(-0.5:0.1:0.5)*ufac"V",
                             ispec=1,
                             solver_kwargs...)

Calculate working electrode current corresponding to rate for species ispec for each voltage in voltages. Returns vector of voltages and vector of currents.