General remarks

All physical quantities are assumed to be consistently represented through their values expressed in basic SI units (m, kg, s, A, K, mol, cd), supported by the LessUnitful.jl package built on top of Unitful.jl.

Electrolyte data

LiquidElectrolytes.AbstractElectrolyteData — Type

abstract type AbstractElectrolyteData

Abstract super type for electrolytes

LiquidElectrolytes.ElectrolyteData — Type

mutable struct ElectrolyteData <: AbstractElectrolyteData

Data for electrolyte. It is defined using Base.@kwdef has keyword constructors like

    ElectrolyteData(nc=3,z=[-1,2,1])

nc::Int64: Number of ionic species.

iϕ::Int64: Potential index in species list.

ip::Int64: Pressure index in species list

D::Vector{Float64}: Mobility coefficient

z::Vector{Int64}: Charge numbers of ions

M0::Float64: Molar weight of solvent

M::Vector{Float64}: Molar weight of ions

v0::Float64: Molar volume of solvent

v::Vector{Float64}: Molar volumes of ions

κ::Vector{Float64}: Solvation numbers

c_bulk::Vector{Float64}: Bulk ion concentrations

ϕ_bulk::Float64: Bulk voltage

p_bulk::Float64: Bulk pressure

Γ_bulk::Int64: Bulk boundary number

ϕ_we::Float64: Working electrode voltage

Γ_we::Int64: Working electrode boundary number

T::Float64: Temperature

RT::Float64: Molar gas constant scaled with temperature

F::Float64: Faraday constant

ε::Float64: Dielectric permittivity of solvent

ε_0::Float64: Dielectric permittivity of vacuum

pscale::Float64: Pressure scaling factor

eneutral::Bool: Local electroneutrality switch

scheme::Symbol: Flux caculation scheme This allows to choose between

:μex (default): excess chemical potential (SEDAN) scheme, see sflux

:act : scheme based on reciprocal activity coefficients, see aflux

:cent : central scheme, see cflux.

epsreg::Float64: Regularization parameter used in rlog

weights::Vector{Float64}: Species weights for norms in solver control.

The default values for electrolyte data are those of an symmetric 0.1M aqueous binary electrolyte at 298.5K with solvation number κ=10, ion molar volumes similar to water molecules and diffusion coefficient 2.0e-9 $m^2/s$. All values given in SI base units.

using LiquidElectrolytes
ElectrolyteData()

<< @example-block not executed in draft mode >>

LiquidElectrolytes.dlcap0 — Method

dlcap0(electrolyte)

Double layer capacitance at zero voltage for symmetric binary electrolyte.

Example

using LessUnitful
ely = ElectrolyteData(c_bulk=fill(0.01ufac"mol/dm^3",2))
round(dlcap0(ely),sigdigits=5) |> u"μF/cm^2"
# output

22.847 μF cm^-2

LiquidElectrolytes.debyelength — Method

debyelength(electrolyte)

Debye length.

using LessUnitful
ely = ElectrolyteData(c_bulk=fill(0.01ufac"mol/dm^3",2))
round(debyelength(ely),sigdigits=5) |> u"nm"
# output

4.3018 nm

LiquidElectrolytes.chemical_potential — Function

    chemical_potential(c, barc, p, v, electrolye)

Calculate chemical potential of species with concentration c

\[ μ = v(p-p_{ref}) + RT\log \frac{c}{\bar c}\]

LiquidElectrolytes.c0_barc — Function

c0_barc(u,electrolyte)

Calculate solvent concentration $c_0$ and summary concentration $\bar c$ from vector of concentrations c using the incompressibility constraint (assuming $κ_0=0$):

\[ \sum_{i=0}^N c_i (v_i + κ_iv_0) =1\]

This gives

\[ c_0v_0=1-\sum_{i=1}^N c_i (v_i+ κ_iv_0)\]

\[c_0= 1/v_0 - \sum_{i=1}^N c_i(\frac{v_i}{v_0}+κ_i)\]

Then we can calculate

\[ \bar c= \sum_{i=0}^N c_i\]

LiquidElectrolytes.rrate — Function

rrate(R0,β,A)

Reaction rate expression

rrate(R0,β,A)=R0*(exp(-β*A) - exp((1-β)*A))

Discretization system

LiquidElectrolytes.PNPSystem — Function

PNPSystem(grid;
         celldata=ElectrolyteData(),
         bcondition=default_bcondition,
         kwargs...)

Create VoronoiFVM system. Input:

grid: discretization grid

celldata: composite struct containing electrolyte data

bcondition: boundary condition

kwargs: Keyword arguments of VoronoiFVM.System

LiquidElectrolytes.pnpunknowns — Function

pnpunknowns(sys)

Return vector of unknowns initialized with bulk data.

LiquidElectrolytes.electrolytedata — Function

electrolytedata(sys)

Extract electrolyte data from system.

LiquidElectrolytes.solventconcentration — Function

   solventconcentration(U::Array, electrolyte)

Calculate vector of solvent concentrations from solution array.

Standard calculations

LiquidElectrolytes.bulkbcondition — Function

bulkbcondition(f,u,bnode,electrolyte)

Bulk boundary condition for electrolyte: set potential, pressure and concentrations to bulk values.

LiquidElectrolytes.dlcapsweep — Function

       dlcapsweep(sys;voltages=(-1:0.1:1)*ufac"V",
                              δ=1.0e-4,
                              molarity=0.1*ufac"mol/dm^3",
                              solver_kwargs...)

Calculate double layer capacitances for voltages given in voltages. Returns vector of voltages and vector of double layer capacitances.

Assumptions:

Only one double layer in the system - close to working electrode

1D domain

LiquidElectrolytes.ivsweep — Function

   ivsweep(sys;
              voltages=(-0.5:0.1:0.5)*ufac"V", solver_kwargs...)

Calculate molar reacton rates and bulk flux rates for each voltage in voltages.

Returns:

transient solution tsol with components t (voltages) and u (solutions)

vector j_we of working electrode molar reaction rates in mol/(m^3*s)

vector j_bulk of bulk electrode in/outflow rates in mol/(m^3*s)

Pre v0.0.21 there was an ispec keyword arg, and the return values were volts, currs, solutions.

Since v0.0.21, the currs vector can be obtained via

currs = [ j[ispec] for j in j_we]