Electrolyte data

LiquidElectrolytes.charge — Function

charge(c,electrolyte)

Calculate charge from vector of concentrations

LiquidElectrolytes.vrel — Function

vrel(ic,electrolyte)

Calculate relative (wrt. solvent) molar volume of i-th species $v_{i,rel}=κ_i+\frac{v_i}{v_0}$.

LiquidElectrolytes.rlog — Function

rlog(u, electrolyte)

Calls rlog(u;eps=electrolyte.epsreg)

rlog(u; eps=1.0e-20)

Regularized logarithm. Smooth linear continuation for x<eps. This means we can calculate a "logarithm" of a small negative number.

LiquidElectrolytes.rexp — Function

rexp(x;trunc=500.0)

Regularized exponential. Linear continuation for x>trunc, returns 1/rexp(-x) for x<trunc.

LiquidElectrolytes.wnorm — Function

wnorm(u,w,p)

Weighted norm with respect to columns

Finite volume system

LiquidElectrolytes.pnpstorage — Function

pnpstorage(f, u, node, electrolyte)

Finite volume storage term

LiquidElectrolytes.pnpreaction — Function

pnpreaction(f, u, node, electrolyte)

Finite volume reaction term

LiquidElectrolytes.default_bcondition — Function

default_bcondition(f,u,bnode,electrolyte)

Default boundary condition amounts to nothing

LiquidElectrolytes.pnpflux — Function

pnpflux(f, u, edge, electrolyte)

Finite volume flux. It calls either sflux, cflux or aflux.

LiquidElectrolytes.sflux — Function

sflux(ic,dϕ,ck,cl,βk,βl,bar_ck,bar_cl,electrolyte)

Sedan flux, see Gaudeul/Fuhrmann 2022

Appearantly first described by Yu, Zhiping and Dutton, Robert, SEDAN III, www-tcad.stanford.edu/tcad/programs/sedan3.html

see also the 198? Fortran code available via https://web.archive.org/web/20210518233152/http://www-tcad.stanford.edu/tcad/programs/oldftpable.html

Verification calculation is in the paper.

LiquidElectrolytes.aflux — Function

aflux(ic,dϕ,ck,cl,βk,βl,bar_ck,bar_cl,electrolyte)

Flux expression based on reciprocal activity coefficents, see Fuhrmann, CPC 2015

LiquidElectrolytes.cflux — Function

aflux(ic,dϕ,ck,cl,βk,βl,bar_ck,bar_cl,electrolyte)

Flux expression based on central differences, see Gaudeul/Fuhrmann 2022

LiquidElectrolytes.dμex — Function

 dμex(βk, βl, electrolyte)

Calculate differences of excess chemical potentials from reciprocal activity coefficients

Electrochemical calculations

LiquidElectrolytes.splitz — Function

splitz(range::AbstractRange)

If range contains zero, split it into two parts, one with values <=0 and one with values >=0. Otherwise, return the range or its reverse, such that first value always is the one with the smallest absolute value.

splitz(range::Vector)

Version of splitz(range::AbstractRange) for vectors.